requirements.txt
--index-url https://download.pytorch.org/whl/cu121
torch==2.4.1
--index-url https://pypi.org/simple
pyyaml==6.0.2
numpy==2.0.0
scikit-learn==1.5.2
lightning==2.4.0
transformers==4.47.1
biopython==1.84
tqdm==4.67.1
fair-esm==2.0.0
matplotlib==3.10.0
rdkit==2024.3.6
networkx==3.4.2
pandas==2.2.3
sentencepiece==0.2.0